+++ src/restfw/tests/test_usage_examples_collector.py
@@ -5,7 +5,7 @@
 """
 import logging
 
-from mountbit.utils.testing import ANY, D
+from cykooz.testing import ANY, D
 from pyramid import httpexceptions
 from pyramid.security import Allow, Everyone, Authenticated, DENY_ALL